+++ website/views/__init__.py
@@ -4,5 +4,4 @@ from .manage_documentation import *
 from .manage_news import *
 from .manage_publications import *
 from .manage_sections import *
-from .manage_workshop import *
 from .pages import *
