+++ workshop/templates/workshop/addeditworkshop.html
@@ -23,7 +23,7 @@
     <br />
     <div class="col-lg-offset-4 col-lg-4 form-actions">
         <button type="submit" class="btn btn-primary" >Submit</button>
-        <a href="{% url 'website:dashboard_workshops' %}"> Cancel </a>
+        <a href="{% url 'workshop:dashboard_workshops' %}"> Cancel </a>
     </div>
 </form>
 {% endif %}
